add Color enum and changed colors in typing_game
