--- enum_level.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class Level(Enum):
-    EASY = 0
-    MEDIUM = 1
-    HARD = 2
